BACKPORT: cpufreq: schedutil: Fix for limits update with fast switch enabled
When the policy limits are applied with fast switch enabled, the
policy->cur is not updated. This can result in incorrect calculation
of the average capacity and any subsequent limit updates.

Update cpufreq_policy_apply_limits_fast() API to return a non-zero
value when the frequency is updated. Make use of this return value
and update policy->cur. While at it, print cpu_frequency trace point,
when frequency is changed due to change in limits.

Change-Id: I51732fa061aac11231d1f18ca70f31f252f0a0dd
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
[Sunneva: We only need cpufreq_policy_apply_limits_fast()
but as we don't have sugov_track_cycles() we must comment
out the bits added to sugov_limits() by this and the
previous patch.]
Signed-off-by: Sunneva N. Mariu <sunnevanattsol@gmail.com>